code size optimization, part 1
rename int32->int16

ring_buffer.h
uart.h
uart.c
main.cpp

Program:    13202 bytes (80.6% full)
(.text, .data, .bootloader)

Data:    568 bytes (55.5% full)
(.data, .bss, .noinit)

+++ src/main.cpp
@@ -29,7 +29,7 @@ void uart_rx_to_tx(void* rx_cb_data)
 	(void)rx_cb_data;
 
 	void *rx_p, *tx_p;
-	uint32_t rx_sz, tx_sz, sz;
+	uint16_t rx_sz, tx_sz, sz;
 	uart_rx_ptr(&rx_p, &rx_sz);
 	uart_tx_ptr(&tx_p, &tx_sz);
 	sz = (rx_sz > tx_sz ? tx_sz : rx_sz);
